# docs/openapi/callpulse-gateways-acl-dialplans.yaml
openapi: 3.0.3
info:
  title: 'FS PBX API v1 — Gateways, Access Controls, Dialplans'
  version: 1.0.0
  description: |
    Contrat des trois ressources ajoutées à l'API v1 de FS PBX pour CallPulse :
    passerelles SIP (`gateways`), listes de contrôle d'accès (`access-controls`)
    et plans de numérotation (`dialplans`).

    Ce document est le contrat de référence : il est écrit avant les contrôleurs
    et fait foi en cas de divergence avec une implémentation.

    ## Authentification

    Toutes les opérations exigent un jeton personnel Sanctum envoyé en
    `Authorization: Bearer <token>`. Deux couches distinctes s'appliquent, dans
    cet ordre, et **elles ne renvoient pas la même forme de corps d'erreur** :

    1. `RequireBearerToken` — en-tête `Authorization` absent → `401`
       `{"success": false, "message": "Bearer token required."}` (forme « middleware »).
    2. `auth:sanctum` — jeton présent mais invalide ou révoqué → `401`
       `{"error": {"type": "authentication_error", ...}}` (forme « enveloppe »).
    3. `AuthorizeUser` — permission manquante → `403` forme « middleware » avec
       `error.code = "forbidden_permission"` ; domaine hors périmètre de
       l'utilisateur → `403` forme « enveloppe » avec `error.code = "forbidden_domain"`.

    Un client doit donc accepter les **deux** formes sur `401` et sur `403`.

    ## Débit

    Limiteur `throttle:api` : **60 requêtes par minute** par utilisateur
...

    ## Codes de retour

    | Code | Signification |
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  '/api/v1/domains/{domain_uuid}/gateways/{gateway_uuid}':
    get:
      tags: [Gateways]
      operationId: getGateway
      summary: 'Lire une passerelle et son état d''enregistrement'
      description: |
        Renvoie la configuration persistée **et** l'état vu par le moteur
        téléphonique, lu sur l'event socket (`sofia xmlstatus gateway`, la même
        source que l'interface d'administration) et rapproché par `gateway_uuid`.

        Si l'event socket est injoignable, `state` vaut `UNKNOWN` et
        `ping_state` / `ping_time_ms` sont `null` : l'absence d'état n'est pas
        une absence de passerelle.

        Permission requise : `gateway_view`.
      parameters:
        - $ref: '#/components/parameters/DomainUuid'
        - $ref: '#/components/parameters/GatewayUuid'
      responses:
        200:
          description: Succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gateway'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }

    patch:
      tags: [Gateways]
      operationId: updateGateway
      summary: 'Modifier partiellement une passerelle'
      description: |
        Modification **partielle** : seuls les champs présents dans le corps sont
        appliqués, les autres conservent leur valeur. Aucun champ n'est requis.

        Si `profile` change, l'ancien **et** le nouveau profil Sofia sont
        rescannés. Envoyer `gateway_acl_cidrs` remplace intégralement la liste
        blanche d'IP de cette passerelle ; envoyer un tableau vide la supprime.

        `password` s'écrit mais ne se relit jamais (voir le schéma `Gateway`).

        Permission requise : `gateway_edit`.
      parameters:
        - $ref: '#/components/parameters/DomainUuid'
        - $ref: '#/components/parameters/GatewayUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatewayUpdateRequest'
            examples:
              desactiver:
                summary: 'Désactiver la passerelle'
                value:
                  enabled: false
      responses:
        200:
          description: 'Passerelle modifiée.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gateway'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }

    delete:
      tags: [Gateways]
      operationId: deleteGateway
      summary: 'Supprimer une passerelle'
      description: |
        Arrête la passerelle sur le moteur (`sofia profile <profile> killgw
        <gateway_uuid>`) **avant** de la supprimer en base, retire les nœuds de
        liste blanche qu'elle possédait, puis recharge les listes d'accès et
        rescanne le profil.

        Réponse **`204 No Content`, sans corps**.

        Permission requise : `gateway_delete`.
      parameters:
        - $ref: '#/components/parameters/DomainUuid'
        - $ref: '#/components/parameters/GatewayUuid'
      responses:
        204:
          description: 'Passerelle supprimée. Aucun corps de réponse.'
        400: { $ref: '#/components/responses/BadRequest' }
        401: { $ref: '#/components/responses/Unauthorized' }
        403: { $ref: '#/components/responses/Forbidden' }
        404: { $ref: '#/components/responses/NotFound' }
        429: { $ref: '#/components/responses/TooManyRequests' }
        500: { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Jeton personnel (Sanctum) d'un utilisateur FS PBX, créé depuis le tableau
        de bord. Un jeton par environnement appelant. L'en-tête
        `Authorization: Bearer <token>` est obligatoire sur toutes les opérations.
  parameters:
    DomainUuid:
      name: domain_uuid
      in: path
      required: true
      description: |
        UUID du domaine (tenant). Doit correspondre à `^[0-9a-fA-F-]{36}$`, sinon
        `400`. Le domaine doit être dans le périmètre de l'utilisateur du jeton,
        sinon `403` (`forbidden_domain`) ; s'il n'existe pas, `404`.
      schema:
        type: string
        format: uuid
      example: 4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b
    GatewayUuid:
      name: gateway_uuid
      in: path
      required: true
      description: |
        UUID de la passerelle. C'est aussi son nom dans Sofia
        (`sofia/gateway/<gateway_uuid>/...`).
      schema:
        type: string
        format: uuid
      example: 47aa96db-3f70-46ce-bb9a-79ccef396f2b
...
  responses:
    BadRequest:
      description: |
        UUID de route malformé, ou corps invalide. **La validation du corps est
        rendue en `400`** (`error.code = "invalid_parameter"`, `error.param` porte
        le premier champ fautif), jamais en `422` : dans cette API, `422` est
        réservé aux erreurs du moteur téléphonique.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            uuid_malforme:
              value:
                error:
                  type: invalid_request_error
                  message: 'Invalid domain UUID.'
                  code: invalid_request
                  param: domain_uuid
                  doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
            corps_invalide:
              value:
                error:
                  type: invalid_request_error
                  message: 'Enter a valid IP address or CIDR range.'
                  code: invalid_parameter
                  param: nodes.0.node_cidr
                  doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
    Unauthorized:
      description: |
        Jeton absent, invalide ou révoqué. **Deux formes de corps** selon la couche
        qui refuse : le middleware `RequireBearerToken` (en-tête absent) renvoie
        `{"success": false, "message": "Bearer token required."}`, tandis qu'un
        jeton invalide passe par l'enveloppe d'erreur standard.
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/ErrorEnvelope'
              - $ref: '#/components/schemas/MiddlewareError'
          examples:
            entete_absente:
              summary: 'Aucun en-tête Authorization'
              value:
                success: false
                message: 'Bearer token required.'
            jeton_invalide:
              summary: 'Jeton présent mais invalide ou révoqué'
              value:
                error:
                  type: authentication_error
                  message: Unauthenticated.
                  code: unauthenticated
    Forbidden:
      description: |
        Permission manquante, ou domaine hors du périmètre de l'utilisateur.
        **Deux formes de corps** : permission manquante → forme « middleware » avec
        `error.code = "forbidden_permission"` et `error.permission` nommant la
        permission attendue ; domaine interdit → enveloppe standard avec
        `error.code = "forbidden_domain"`.
      content:
        application/json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/MiddlewareError'
              - $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            permission_manquante:
              value:
                success: false
                message: 'Forbidden (missing permission).'
                error:
                  code: forbidden_permission
                  permission: gateway_add
            domaine_interdit:
              value:
                error:
                  type: invalid_request_error
                  message: 'You do not have access to this domain.'
                  code: forbidden_domain
                  param: domain_uuid
                  doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
    NotFound:
      description: |
        Ressource inexistante, ou existante mais rattachée à un autre domaine que
        celui de la route.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              type: invalid_request_error
              message: 'Gateway not found.'
              code: resource_missing
              param: gateway_uuid
              doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
...
    TooManyRequests:
      description: |
        Débit dépassé : 60 requêtes par minute et par utilisateur authentifié
        (à défaut, par adresse IP). Les en-têtes `Retry-After` et
        `X-RateLimit-Remaining` accompagnent la réponse.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: |
        Erreur interne. Le message est volontairement générique ; le détail est
        journalisé côté serveur.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              type: api_error
              message: 'Internal server error.'
              code: internal_error
              doc_url: 'https://www.fspbx.com/docs/api/v1/errors/'
  schemas:
...
    ErrorEnvelope:
      type: object
      description: "Enveloppe d'erreur standard de l'API v1."
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              description: "Famille d'erreur."
              enum:
                - invalid_request_error
                - authentication_error
                - not_found
                - forbidden_method
                - api_error
            message:
              type: string
              description: "Message lisible. Pour un `422`, il contient la réponse brute du moteur téléphonique."
            code:
              type: string
              nullable: true
              description: 'Code machine, stable — c''est lui qu''un client doit tester.'
              enum:
                - unauthenticated
                - invalid_request
                - invalid_parameter
                - forbidden
                - forbidden_domain
                - forbidden_permission
                - resource_missing
                - resource_not_found
                - switch_error
                - internal_error
            param:
              type: string
              nullable: true
              description: 'Champ ou paramètre fautif, quand il est identifiable.'
            doc_url:
              type: string
              nullable: true
    MiddlewareError:
      type: object
      description: |
        Forme d'erreur renvoyée par les middlewares d'authentification et
        d'autorisation, **distincte** de l'enveloppe standard.
      required: [success, message]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          example: 'Forbidden (missing permission).'
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
              enum: [forbidden_permission]
            permission:
              type: string
              description: 'Permission attendue.'
              example: gateway_add

    # --- Passerelles ---------------------------------------------------------
    GatewayState:
      type: string
      description: |
        État de la passerelle tel que rapporté par le moteur téléphonique.

        **L'énumération est fermée** : tout autre état renvoyé par FreeSWITCH
        (`TRYING`, `REG_FAILED`, …) est ramené à `UNKNOWN`. Un client peut donc
        traiter cette valeur comme un ensemble fini.

        - `REGED` — enregistrée auprès du proxy distant ;
        - `UNREGED` — désenregistrée ;
        - `NOREG` — passerelle configurée sans enregistrement (`register = false`) ;
        - `FAIL_WAIT` — échec d'enregistrement, attente avant nouvelle tentative ;
        - `DOWN` — passerelle arrêtée sur le moteur ;
        - `UNKNOWN` — état indéterminé, **notamment quand l'event socket est
          injoignable** : ce n'est pas un état de la passerelle mais une absence
          d'information.
      enum: [REGED, NOREG, FAIL_WAIT, DOWN, UNREGED, UNKNOWN]
      example: REGED
...
    GatewayUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/GatewayCommon'
        - type: object
          description: |
            Modification partielle : aucun champ n'est requis, seuls les champs
            présents sont appliqués.
    Gateway:
      type: object
      description: |
        Représentation d'une passerelle en réponse.

        **Le mot de passe SIP (`password`) n'est jamais renvoyé, sur aucune
        opération** — ni à la création, ni en lecture, ni après modification. Il
        s'écrit et ne se relit pas. Un client qui a besoin de le vérifier doit le
        réécrire, pas le lire.
      required: [gateway_uuid, object, domain_uuid, gateway, state]
      properties:
        gateway_uuid:
          type: string
          format: uuid
          description: 'Identifiant de la passerelle, et son nom dans Sofia.'
          example: 47aa96db-3f70-46ce-bb9a-79ccef396f2b
        object:
          type: string
          enum: [gateway]
        domain_uuid:
          type: string
          format: uuid
        gateway:
          type: string
          example: client_poste_250
        username:
          type: string
          nullable: true
        proxy:
          type: string
          nullable: true
        register:
          type: boolean
        register_transport:
          type: string
          nullable: true
          enum: [udp, tcp, tls]
        from_user:
          type: string
          nullable: true
        from_domain:
          type: string
          nullable: true
        expire_seconds:
          type: integer
          nullable: true
        retry_seconds:
          type: integer
          nullable: true
        ping:
          type: integer
          nullable: true
        channels:
          type: integer
          nullable: true
        context:
          type: string
          nullable: true
        profile:
          type: string
          nullable: true
          example: external
        hostname:
          type: string
          nullable: true
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
        state:
          $ref: '#/components/schemas/GatewayState'
        ping_state:
          type: string
          nullable: true
          description: |
            État de la sonde SIP OPTIONS, tel que rapporté par le moteur.
            `null` si `ping` n'est pas configuré sur la passerelle, ou si le
            moteur est injoignable.
          example: UP
        ping_time_ms:
          type: number
          nullable: true
          description: 'Temps de sonde en millisecondes. `null` si indisponible.'
          example: 12.4
        switch_response:
          type: string
          nullable: true
          description: |
            Réponse brute du moteur à la commande de démarrage. **Une valeur
            commençant par `-ERR` signifie que la base est à jour mais que
            FreeSWITCH n'a pas pris la passerelle en compte.**
          example: '+OK'
        acl_response:
          type: string
          nullable: true
          description: 'Réponse brute du moteur au rechargement des listes d''accès (`reloadacl`). Mêmes règles de lecture que `switch_response`.'
...